Combined jewellery, charges, and supplies box into one box, removed abstract box. cleaned up code.
